feat: assertion limit, mailer disable based on env; refactor: move db config to seperate folder
