+++ backend/src/modules.d.ts
@@ -11,5 +11,9 @@ declare namespace NodeJS {
         S3_PUBLIC_URL: string;
         SES_SENDER_NAME: string;
         SES_SENDER_ADDRESS: string;
+        MAIL_DISABLE: string;
+        MAX_REQ_PER_SECOND: string;
+        MAX_AWARD_PER_BADGE: string;
+        REDIS_URI: string;
     }
 }
